Return latest version if requested version is any
I believe this is a bug in semver crate but according to NPM people, and
@steveklabnik; a version like "0.1.0-pre.0" or "0.1.0-beta.0" *is not a
real version* and any version that contains a pre-release string is getting
different treatment by semver.

IMO this is utterly wrong. Docs.rs is probably the first victim of this
behavior and it will affect more in future. `*` means any and it should
match any version. It doesn't matter if a version have a pre-release string
or not. Even older versions of node's semver were matching "*" for any
version. I don't know why they changed this behavior but I am not surprised
any means "anything except a version that contains pre-release" in JS
world.

Fixes: #80
Ref: https://github.com/steveklabnik/semver/issues/98