Try to build crates if there is any in queue
In addition to #90 this change will also run `build_queue` in it's own
thread to catch panics (only panicked twice in last 6 months).

This commit also introduces an attempt count to avoid endless tries to build
yanked crates. This only happens if a crate gets yanked before docs.rs tries.
docs.rs will only try to build a crate 5 times and ignore them after it but it
will still leave it in queue for inspection.