Adding a configuration option to specify rustc version used
This code is maybe not as elegant as it could be but should hopefully work? I'm
not 100% sure how the crates are built (i.e. how the working directories are
built and scoped) so I hope that `rustup` is invoked in the correct directory.

If that's not the case, then this can be changed.

Additionally I am aware that this is a **work in progress** and further changes
are required on docs.rs and the compilers part to make "hot-swapping" versions
easier. See #225 for more details. This PR addresses #228