Don't override the user's fonts
This patch is similar in spirit to rust-lang/blog.rust-lang.org#263,
though it's also less invasive; it removes the specific font overrides
in style.css so that the browser will use the user's configured fonts.

The net result of the change is that the page will look less "uniform"
across different browsers, OSes, and devices, but the overall aesthetic
should be preserved. It also means that user preferences are respected,
and that legibility is improved for users with particular font needs, or
on platforms that sometimes gets confused about how to render fonts
nicely (ehrm, Linux).