Fix uploads of large crates to S3 breaking the client
Currently, we reuse the client for all files that are uploaded per call
to the relevant code. Unfortunately, S3 will prevent connections from
being long-lived (it's unknown what the exact timeout is, but we hit it
relatively rarely). This code attempts to retry S3 uploads up to 3 times
while replacing the client each time.