rustwide_builder: use a temp dir to store docs not yet uploaded
Before uploading them to the database/S3, the documentation builder
needs to copy the generated files from the target directory (which will
be purged) to another local directory.

That directory is not needed after the whole crate is built, but the
code after the rustwide refactory didn't delete it. This caused an
outage due to the disk being full.

This commit replaces the local directory inside the prefix with
temporary directories removed when the build finishes, which both fixes
the problem and also reduces the uses of the prefix.