web: show the priority in the queue
One of the tools docs.rs operators have to ensure a fair build queue to
everyone is to manually deprioritize slow-to-build crates. Before this
the priority wasn't exposed anywhere in the UI though, making the queue
look weird when a bunch of crates were deprioritized.

This changes the queue page to show the priority in the build queue if
it's not '0'.