Extend size of `:target`ed elements
When the browser follows a link with a `#fragment`, it scrolls up or
down to put the element with the matching `id` attribute at the top
of the window. In rustdoc output, that's usually section headers.

docs.rs's floating top navbar hides that element. This PR extends
the height of the element the browser is scrolling to, in a way that
doesn't affect its visual appearance, so that it lands below the
navbar and is visible to the user.