use the latest rustdoc css for the whole site
This commit removes the vendored CSS of Rustdoc 1.10.0-nightly, changing
the site to always use the lastest Rustdoc's one. To do that, the Docker
entrypoint is changed to add the essential files the first time the
container is started, and the frontend now queries the database to see
the version of the files to include.

The version is cached by the first page loaded by users, and doesn't
change unless the application is restarted. This is fine, since the
application is restarted often enough and using a CSS older by a few
nightlies is not so big of a deal. The cache also prevents response time
regressions.