Allow navigating back to default target
The code as of the previous commit would let you navigate away from the
default, but not navigate back. This adds the default target to
`successful_targets` to add a link to the dropdown in 'Platform'.

After adding that, the link would give a 404
(because we treat paths literally when sending them to the database),
so the metadata for every crate now has to include the default target,
which allows redirecting to /:crate/:version/:module/ when visiting
/:crate/:version/:module/:default-target