CrateDetails: add functions to work with latest version; refactor rustdoc
This commit manages to remove _a lot_ of code from src::web::mod. The removed code was intended to determine the latest version of a list of versions. Its only caller, rustdoc, passed a list of versions directly from CrateDetails.
This could be simplified by moving this logic into CrateDetails, which already has a sorted list of versions.struct.