update CrateDetails tests to handle yanked crates correctly
When a crate is yanked from crates.io, its documentation might not yet
be removed from docs.rs. So it is possible to create CrateDetails of a
release that was build successfully but yanked afterwards. Since this
release was built successfully, it will not have a reference to a last
successful build. But since it was yanked, other release will not point
to this release as last successful build.