Override entrypoint of minio to make 'rust-docs-rs' bucket #463
docs.rs expects there to be a bucket 'rust-docs-rs', the minio Docker
image does not provide a nice way to launch with a default bucket.

As a workaround, we have to override the entrypoint to do some work
before starting the minio server. When the minio server is launched
with a directory that is not empty, it will interpret the directories
present as buckets it needs to create. So by adding an empty directory
'rust-docs-rs', we can trick minio in creating a similarly named
bucket.

This solution is a mixture of the following comments:
 - https://github.com/minio/minio/issues/4769#issuecomment-320223201
 - https://github.com/minio/minio/issues/4769#issuecomment-320319655