Allow users to copy dependency text.
This change adds a clipboard icon next to the crate name that
copies the crate information in the format required for
the user's TOML file. For example, a user on the page
for the `regex` crate at version `1.3.0` will have
`regex = 1.3.0` copied to their clipboard.

Currently, the user experience for this is fairly bare-bones.
It may be useful to have some kind of notification from the UI
that the copy happened, such as using a tooltip or check mark.