switch to empty-library to build essential files
Before this commit docs.rs was using the acme-client 0.0.0 crate to
generate the essential files for a nightly, but that crate actually
contained some non-default values in it, which caused a breakage on
2020-06-12.

This commit switches that crate for empty-library 1.0.0, an empty
library crate with the bare minimum information accepted by crates.io
and with literally no code in it. The crate was just published by me.