remove daemonization code from docs.rs
This commit removes the forking code in the docs.rs daemon, forcing it
to always run in the background. The rationale for the change is that
forking adds complexity and sets extra constraint for development, while
it's not needed anymore on modern infrastructure.

The --foreground flag is kept in the binary for compatibility, and it
will only print a warning. Once this commit is deployed in production
for a while, the flag can be removed.