docbuilder: remove --skip-if-log-exists
As far as I'm aware this flag was never used since the Rust project
started to take care of docs.rs, and it's not much useful.

Basically, every time a build happened the DocBuilder recorded the crate
name and version in a text file, and the flag prevented a build if the
related names and versions were present in such file. Checking if the
build is successful in the database is more reliable.