Make crate menu onclick rather than onhover.
This makes for a more consistent experience on mobile and desktop.

It also, in my opinion, makes for better usability. It's always clear
how to close a menu - the same way as you opened it, by clicking. It
means that menus aren't opened by stray mouse movements. There are more
details on the usability of hover menus at
https://uxmovement.com/navigation/why-hover-menus-do-users-more-harm-than-good/.