menus: Tweak JS handling of clicks.
Prior to this change, if you clicked a menu to open it, then clicked to
close it, the menu heading would still have a grey background
highlighting it. That's due to a CSS rule in pure-menu that causes
focused elements to have a highlighted color. To fix: unfocus the menu
when it is closed via mouse (but leave it focused when it is closed via
keyboard).

Remove a mouseover handler that is not needed anymore.