Fix tabbing through the giant crate menu
Without this tabindex attribute, the scroll view itself gets selected.
With this attribute added, focus jumps straight to the menu item within,
which is what we want, since it contains nothing but menu items, and
most of the keyboard shortcuts you would use to interact with the scroll
view are trapped by JavaScript anyway.