Remove unused `date_added` columns from the database
This fixes the following panic that only appeared in production:

```
Error: Failed to run database migrations

Caused by:
    Error running migration migrate timestamp to be timezone aware, error: db error: ERROR: column "date_added" does not exist
```

This is probably because the original schema in the migrations file
didn't match the schema in the production database (the schema
for migrations was written later). Fortunately, the `date_added` column
is completely unused, so it can simply be deleted locally. This should
not impact the production database, except for allowing the TIMESTAMP
migration to go through.