Remove `database_file_handler`; give precendence to router error messages
The database handler was unused; the router already serves the favicon and it
doesn't do anything else. Note that 'database handler' is different
from 'files served from storage', the router already handles those.

This also fixes the long standing bug that all 404 errors are either
'crate not found' or 'resource not found'.