replace `git2` with `gitoxide` in build-script
Note that the implementation takes liberties as it assumes what's
intended is a unique short id of the commit pointed to by HEAD, instead
of hardcoded 7 hex characters. `shorten()` respects `core.abbrev`
and will automatically use more characters if the chosen hash isn't
unique or not unique enough relative to the size of the git repository.

This is exactly what git does when asking for a short hash.