upgrade crates-index-diff to v13, with 'normalization' support (#1807)
* upgrade crates-index-diff to v11, with 'normalization' support

See this PR for details: https://github.com/rust-lang/crates.io/pull/5066

* Use latest version of crates-index-diff for bugfixes

* Use fixed version of crates-index-diff to allow git2 fetches to succeed

* fix build after merge

- lock file broke due to bad merge
- git v0.15 isn't supported yet and the latest fix switches back to git2

* update `crates-index-diff` to latest version

* upgrade rustwide to latest version

* fix compile errors

* Upgrade `crates-index-diff` to latest release.

* Don't use `git::oid` in favor of copying `git::ObjectId`.

For consistency and simplicity, as it also removes some noise.

* remove git2 from dependencies

* replace `git2` with `gitoxide` in build-script

Note that the implementation takes liberties as it assumes what's
intended is a unique short id of the commit pointed to by HEAD, instead
of hardcoded 7 hex characters. `shorten()` respects `core.abbrev`
and will automatically use more characters if the chosen hash isn't
unique or not unique enough relative to the size of the git repository.

This is exactly what git does when asking for a short hash.

* simplify access to `config.json` file at a given reference

* Update `crates-index-diff` for minor fix (wrong type name)

* upgrade `crates-index-diff` and `git-repository` to latest version

This improves diff performance, but mostly just makes it compile as
v0.25 of git-repository is broken.