Remove footer (#2013)
Move "About" and "Privacy policy" into the Rust menu.

The footer required a lot of workarounds to work with the various bits of
HTML that rustdoc (and doc authors!) generated. It's hard enough to make
the topbar work consistently; adding a footer to nearly-arbitrary HTML
is too much, particularly when we can achieve the same thing in the
topbar.