Set X-Robots-Tag: noindex for nonlatest rustdoc
We want /latest/ URLs to be indexed, but not /1.2.3/ URLs, because the
latter compete for pagerank with /latest/ and aren't usually what people
want.

This replaces the previous canonical URL header. Canonical URLs were
doing some good, but did not work in all cases. In particular, if some
older version of a crate's documentation was very different than the
latest version, Google would not accept the canonicalization. This would
sometimes result in the old version still showing up in the search
results instead of /latest/.