Rely on cargo adding features for optional dependencies
Since the `dep:` syntax was stabilized Cargo now reports optional
dependencies as features for `cargo metadata`, so we don't need to
manually add these ourselves, and it takes into account whether they've
been hidden via `dep:` which we didn't.

We don't get told which features are from optional dependencies, which
is something we have been tracking in the database. But we haven't
started using this data in the UI in the 2 years it's been tracked so we
may as well drop it.

The `serde` test checks that we don't regress on crates that expose
their optional dependencies, the `stylish-core` test did not originally
pass and checks that we now strip out optional dependencies hidden
behind `dep:`.