Update the favicon with a sharper, tuned version.
The new version keeps the style of the previous one uses the R from the Rust logo. It has been made slightly taller to be more clear at small resolution and given a clearer border for stronger constrast on varying backgrounds.

It contains 16x16 and 32x32 versions, rendered to show well on Retina as well non-Retina screens.