Allow index-update during lockfile regeneration
Since the default index changed to the sparse-index we have had issues with
regenerating the lockfile if the dependency list changed, we relied on the
previous fetch having pulled down details for all crates to use, but now it's
only guaranteed to have details on the crates in the original locked dependency
list. If new versions of crates (like log 0.4.17 -> 0.4.19) changed what
dependencies it used we wouldn't have the details on these dependencies to use
the updated version in the lockfile.