Mark default features with a `*`
Fixes #2531 by marking every (transitive) default feature with a `*`.
The features list is already sorted according to depth-first `"default"` order first, and according to the number of sub-features.

I changed that so the secondary sort order is alphabetic, as that might make more sense than sorting by number of sub-features.