web/error: add `JsonAxumNope` for JSON APIs
This aims to satisfy these wishes:

- Use the same error enum (`AxumNope`) for both handlers returning
  HTML or JSON.

- Yet allow those handlers to specify whether HTML or JSON should be
  returned.

- Not change the exising code returning `AxumNope`--still convert to
  HTML by default.

Because `AxumNope` also contains a case (`Search`) that is fixed to
producing HTML, but shouldn't ever be used by handlers returning JSON,
that works with a runtime error in case the latter assumption isn't
being followed.

The approach is to add a new wrapper around `AxumNope` called
`JsonAxumNope` and matching type def `JsonAxumResult`. The wrapper
also implements `IntoResponse`, but produces JSON. Endpoints wishing
to return JSON errors need to use that wrapper and specify
`JsonAxumResult` as their return type.