Properly link licenses on crate pages
Fixes #2738

I'm not happy with the styling but I simply don't understand the CSS here well enough to do better without putting in significant effort.

I decided to consistently link to SPDX instead of also trying to link to https://choosealicense.com/ likes crates.io, but it's an easy change to make if we want to add that too.